semi-final version of api part

+++ src/main/java/pl/springboot/bookrentalservice/manager/BookManager.java
@@ -12,6 +12,8 @@ import pl.springboot.bookrentalservice.dao.entity.RentalService;
 import pl.springboot.bookrentalservice.dao.modelWrappers.SearchWrapper;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -21,11 +23,13 @@ import java.util.stream.StreamSupport;
 public class BookManager {
     private BookRepo bookRepo;
     private BookRentalRepo bookRentalRepo;
+    private BookRentalManager bookRentalManager;
 
     @Autowired
-    public BookManager(BookRepo bookRepo, BookRentalRepo bookRentalRepo) {
+    public BookManager(BookRepo bookRepo, BookRentalRepo bookRentalRepo, BookRentalManager bookRentalManager) {
         this.bookRepo = bookRepo;
         this.bookRentalRepo = bookRentalRepo;
+        this.bookRentalManager = bookRentalManager;
     }
 
     public Optional<Book> findById(Long id) {
@@ -36,30 +40,79 @@ public class BookManager {
         return bookRepo.findAll();
     }
 
-    public Book save(Book videoCassette) {
-        bookRentalRepo.save(new RentalBook(0L, videoCassette.getId(),false));
-        return bookRepo.save(videoCassette);
+    public Object save(Book book) {
+        if(book.getId()==null || book.getId()==0){
+            bookRepo.save(book);
+            bookRentalRepo.save(new RentalBook(0L, book.getId(),false));
+            return book;
+        }
+     return "400";
     }
 
-    public void deleteById(Long id) {
-        bookRepo.deleteById(id);
+    public Object deleteById(Long id) {
+        Optional<RentalBook> rentalBook = bookRentalManager.findByIdBook(id);
+
+        if(rentalBook.isPresent()) {
+            bookRentalRepo.delete(rentalBook.get());
+            bookRepo.deleteById(id);
+            return 200;
+        }
+        return 400;
     }
 
+    boolean tryParseInt(String value) {
+        try {
+            Integer.parseInt(value);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
 
     public Iterable<Book> search(SearchWrapper searchWrapper) {
 
         Iterable<Book> books = bookRepo.findAll();
         Stream<Book> bookStream = StreamSupport.stream(books.spliterator(),false);
-        ;
+        String[] params = searchWrapper.getDate().split("-");
+        LocalDate localDate=null;
+        ArrayList<Integer> dateParams = new ArrayList<Integer>();
+
+
+        for (String param:params) {
+            if(!tryParseInt(param)){
+             searchWrapper.setDate(null);
+            }
+            else{
+                dateParams.add(Integer.parseInt(param));
+            }
+        }
+        if(dateParams.size() !=3)
+        {
+            searchWrapper.setDate(null);
+        }
+        else
+         localDate =  LocalDate.of(dateParams.get(0),dateParams.get(1),dateParams.get(2));
+
         if (searchWrapper.getAuthor() != null)
-        bookStream = bookStream.filter(x -> x.getAuthor().equals(searchWrapper.getAuthor()));
+        bookStream = bookStream.filter(x -> x.getAuthor().contains(searchWrapper.getAuthor()));
         if (searchWrapper.getTitle() != null)
-        bookStream = bookStream.filter(x -> x.getTitle().equals(searchWrapper.getTitle()));
-        if (searchWrapper.getDate() != null)
-        bookStream = bookStream.filter(x -> x.getProductionYear()==(searchWrapper.getDate()));
+        bookStream = bookStream.filter(x -> x.getTitle().contains(searchWrapper.getTitle()));
+        if (localDate != null) {
+            LocalDate finalLocalDate = localDate;
+            bookStream = bookStream.filter(x -> x.getProductionYear().equals(finalLocalDate));
+        }
 
 
         return bookStream.collect(Collectors.toList());
+    }
+
+    public Object update(Book book) {
+        if(book.getId() != null && book.getId()>0)
+        {
+            bookRepo.save(book);
+            return book;
+        }
+        return "400";
     }
 
 //    @EventListener(ApplicationReadyEvent.class)
